setup gradle and inputs for granular tests
