--- src/testInputs/java/com/example/simple_type_names/ValueC.java
@@ -1,4 +1,0 @@
-package com.example.simple_type_names;
-
-public class ValueC {
-}
