+++ src/testInputs/java/com/example/simple_type_names/named/Config.java
@@ -1,0 +1,4 @@
+package com.example.simple_type_names.named;
+
+public class Config {
+}
