+++ src/testInputs/java/com/example/simple_type_names/named/ValueD.java
@@ -1,4 +1,4 @@
-package com.example.simple_type_names;
+package com.example.simple_type_names.named;
 
 public class ValueD {
 	public static void value(ValueD valueD) {
